refactor(era5): move daily ERA5 processing to core_daily
